separating _utils.py into _input_utils.py and _output_utils.py
