+++ src/pymatgen/io/jdftx/_output_utils.py
@@ -136,34 +136,6 @@ def multi_getattr(varbase: Any, varname: str):
     return varbase
 
 
-def flatten_list(tag: str, list_of_lists: list[Any]) -> list[Any]:
-    """Flatten list of lists into a single list, then stop.
-
-    Flatten list of lists into a single list, then stop.
-
-    Parameters
-    ----------
-    tag : str
-        The tag to flatten the list of lists for.
-    list_of_lists : list[Any]
-        The list of lists to flatten.
-
-    Returns
-    -------
-    list[Any]
-        The flattened list.
-    """
-    if not isinstance(list_of_lists, list):
-        raise TypeError(f"{tag}: You must provide a list to flatten_list()!")
-    flist = []
-    for v in list_of_lists:
-        if isinstance(v, list):
-            flist.extend(flatten_list(tag, v))
-        else:
-            flist.append(v)
-    return flist
-
-
 def _brkt_list_of_3_to_nparray(line: str) -> np.ndarray:
     """Return 3x1 numpy array.
