+++ src/pymatgen/io/jdftx/jdftxoutfileslice.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
 from pymatgen.core.periodic_table import Element
 from pymatgen.core.trajectory import Trajectory
 from pymatgen.core.units import Ha_to_eV, ang_to_bohr
-from pymatgen.io.jdftx._utils import (
+from pymatgen.io.jdftx._output_utils import (
     find_all_key,
     find_first_range_key,
     find_key,
