+++ src/pymatgen/io/jdftx/outputs.py
@@ -15,7 +15,7 @@ import pprint
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
-from pymatgen.io.jdftx._utils import read_outfile_slices
+from pymatgen.io.jdftx._output_utils import read_outfile_slices
 from pymatgen.io.jdftx.jdftxoutfileslice import JDFTXOutfileSlice
 
 if TYPE_CHECKING:
